Renamed SpringBootVersionHelper to SpringBootHelper and folded DependencyHelper into it.
